requirements.txt
fastapi
pydantic
python-dotenv
langchain
langgraph
langchain-google-genai
langchain-tavily
langsmith
google-generativeai
uvicorn